Migrate plugin code to Flutters V2 embedding
